refactor: rename drizzleTypes.ts to bindTypes.ts
