+++ src/index.ts
@@ -1,6 +1,6 @@
 import { Hono } from 'hono'
 import usersRouters from './routers/users.router'
-import type { Env } from './types/drizzleTypes'
+import type { Env } from './types/bindTypes'
 
 const app = new Hono<{ Bindings: Env }>()
