+++ src/types/bindTypes.ts
@@ -1,0 +1,4 @@
+export type Env = {
+	DB: D1Database
+	KV: KVNamespace
+}
