--- src/types/drizzleTypes.ts
@@ -1,3 +1,0 @@
-export type Env = {
-    DB: D1Database
-  }
